migrate  import lithology csv from Flask to Fire script
