--- app/scripts/foo.py
@@ -1,7 +1,0 @@
-import click
-
-from app import app
-
-@app.cli.command()
-def historical_records():
-    print 'hi'
