--- app/setup.py
@@ -1,24 +1,0 @@
-import os
-from setuptools import setup
-
-from dotenv import load_dotenv
-
-
-# load_dotenv(".env", verbose=True)
-# path = os.environ.get("PASSENGER_BASE_PATH")
-# sys.path.append(path)
-
-# cwd = os.getcwd()
-# print('--------->>')
-# print(path)
-# print(cwd)
-# print('--------->>')
-
-# setup(
-#     name='flask-my-extension',
-#     entry_points={
-#         'console_scripts': [
-#             'console=console:test'
-#         ],
-#     },
-# )
